Split project into Multi modules
OkHttp is now in a separate module.
MTOM is dependant on OkHttp for now, will need to refactor this later.
JDK module is only dependant on api
