+++ api/src/main/java/no/storebrand/shampoo/FromElement.java
