+++ api/src/main/java/no/storebrand/shampoo/JDOM2Utils.java
