+++ api/src/main/java/no/storebrand/shampoo/Result.java
