+++ api/src/main/java/no/storebrand/shampoo/SoapClient.java
@@ -1,0 +1,5 @@
+package no.storebrand.shampoo;
+
+public interface SoapClient {
+    Result<SoapFault, SoapDocument> execute(SoapRequest req);
+}
