+++ api/src/main/java/no/storebrand/shampoo/SoapFault.java
