+++ api/src/main/java/no/storebrand/shampoo/SoapHeader.java
