--- src/main/java/no/storebrand/shampoo/SoapRequest.java
@@ -1,31 +1,0 @@
-package no.storebrand.shampoo;
-
-import okhttp3.Headers;
-
-import java.util.List;
-
-public final class SoapRequest {
-    public final Headers headers;
-    public final SoapDocument soapDocument;
-    public final SoapAction action;
-
-    private SoapRequest(Headers headers, SoapDocument document, SoapAction action) {
-        this.headers = headers;
-        this.action = action;
-        this.soapDocument = document;
-    }
-
-    public static SoapRequest soap11(SoapBody body, List<SoapHeader> header, SoapAction action) {
-        SoapDocument doc = SoapDocument.soap11(header, body);
-        return new SoapRequest(Headers.of(), doc, action);
-    }
-
-    public static SoapRequest soap12(SoapBody body, List<SoapHeader> header, SoapAction action) {
-        SoapDocument doc = SoapDocument.soap12(header, body);
-        return new SoapRequest(Headers.of(), doc, action);
-    }
-
-    public SoapRequest addHttpHeader(String name, String value) {
-        return new SoapRequest(headers.newBuilder().add(name, value).build(), soapDocument, action);
-    }
-}
